requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.26.0
importlib-metadata==6.0.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.15.0
